moved water configuration to new mdpow.forcefields
