+++ mdpow/tables.py
@@ -8,79 +8,14 @@ Tables of hard-coded values used in mdpow
 =========================================
 
 TODO: Move these data into files in the top directory.
+
+.. SeeAlso:: :mod:`mdpow.forcefields`
 """
 
 from __future__ import absolute_import
 
 from numkit.observables import QuantityWithError
 
-import logging
-logger = logging.getLogger("mdpow.tables")
-
-#------------------------------------------------------------
-# Gromacs water models
-#------------------------------------------------------------
-
-#: See the file ``top/oplsaa.ff/watermodels.dat`` for a description of
-#: available water models that are bundled with MDPOW.
-GMX_WATERMODELS_DAT="""
-tip4p   TIP4P  TIP 4-point, recommended
-tip3p   TIP3P  TIP 3-point
-tip5p   TIP5P  TIP 5-point
-spc     SPC    simple point charge
-spce    SPC/E  extended simple point charge
-"""
-
-class GromacsWaterModel(object):
-    """Data for a water model."""
-    def __init__(self, identifier, name=None, itp=None, coordinates=None,
-                 description=None,
-                 forcefield="OPLS-AA"):
-        self.identifier = identifier
-        self.name = name if name is not None else str(identifier).upper()
-        self.itp = itp if itp is not None else self.guess_filename('itp')
-        self.coordinates = coordinates if coordinates is not None else self.guess_filename('gro')
-        self.description = description
-        self.forcefield = forcefield
-
-    def guess_filename(self, extension):
-        return self.identifier.lower() + '.' + str(extension)
-
-    def __repr__(self):
-        return "<{0[name]} water: identifier={0[identifier]}, ff={0[forcefield]}>".format(vars(self))
-
-def _create_water_models(watermodelsdat):
-    models = {}
-    for line in GMX_WATERMODELS_DAT.split('\n'):
-        line = line.strip()
-        if not line or line.startswith('#'):
-            continue
-        fields = line.split()
-        identifier, name = fields[:2]
-        description = " ".join(fields[2:])
-        models[identifier] = GromacsWaterModel(identifier, name=name, description=description)
-    return models
-
-#: Use the default water model unless another water model is chosen in the runinput
-#" file (``setup.watermodel``).
-DEFAULT_WATER_MODEL = "tip4p"
-
-#: Dictionary of :class:`GromacsWaterModel` instances, one for each Gromacs water
-#: model  available under the force field directory. The keys are the water model
-#: identifiers.
-#: For OPLS-AA the following ones are available.
-GROMACS_WATER_MODELS = _create_water_models(GMX_WATERMODELS_DAT)
-
-def get_water_model(watermodel=DEFAULT_WATER_MODEL):
-    """Return a :class:`GromacsWaterModel` corresponding to identifier *watermodel*"""
-
-    try:
-        return GROMACS_WATER_MODELS[watermodel]
-    except KeyError:
-        msg = "{0} is not a valid water model: choose one from {1}".format(
-            watermodel, ", ".join(GROMACS_WATER_MODELS.keys()))
-        logger.error(msg)
-        raise ValueError(msg)
 
 #------------------------------------------------------------
 # constants
